add use of env variables, update BE structure
